refactor(nat): move to internal
