+++ cmd/msg/main.go
@@ -13,7 +13,7 @@ import (
 	"nimona.io/pkg/eventbus"
 	"nimona.io/pkg/exchange"
 	"nimona.io/pkg/keychain"
-	"nimona.io/pkg/nat"
+	"nimona.io/internal/nat"
 	"nimona.io/internal/net"
 	"nimona.io/pkg/object"
 	"nimona.io/pkg/peer"
