+++ internal/daemon/daemon.go
@@ -9,14 +9,14 @@ import (
 	_ "github.com/mattn/go-sqlite3"
 
 	"nimona.io/internal/daemon/config"
+	"nimona.io/internal/nat"
+	"nimona.io/internal/net"
 	"nimona.io/pkg/context"
 	"nimona.io/pkg/crypto"
 	"nimona.io/pkg/errors"
 	"nimona.io/pkg/eventbus"
 	"nimona.io/pkg/exchange"
 	"nimona.io/pkg/keychain"
-	"nimona.io/pkg/nat"
-	"nimona.io/internal/net"
 	"nimona.io/pkg/object"
 	"nimona.io/pkg/objectmanager"
 	"nimona.io/pkg/peer"
